[Cleanup] Split functions from Misc.cpp/.h
- Rules calculation functions
- PortStatus functions
- String Parser functions
