+++ src/src/Commands/Rules.cpp
@@ -12,6 +12,7 @@
 #include "../Globals/Settings.h"
 
 #include "../Helpers/Misc.h"
+#include "../Helpers/Rules_calculate.h"
 #include "../Helpers/StringConverter.h"
 
 String Command_Rules_Execute(struct EventStruct *event, const char *Line)
