+++ src/src/Globals/Cache.h
@@ -1,9 +1,10 @@
 #ifndef GLOBALS_CACHE_H
 #define GLOBALS_CACHE_H
 
+#include "../DataStructs/Caches.h"
+
 void clearAllCaches();
 
-struct Caches;
 extern Caches Cache;
 
 #endif // GLOBALS_CACHE_H
