+++ src/src/Helpers/StringConverter.cpp
@@ -16,6 +16,8 @@
 #include "../Globals/Plugins.h"
 #include "../Globals/Settings.h"
 
+#include "../Helpers/StringParser.h"
+
 #include "Misc.h"
 
 #include "../../ESPEasy_Log.h"
