+++ src/src/PluginStructs/P023_data_struct.cpp
@@ -1,6 +1,7 @@
 #include "P023_data_struct.h"
 
 #include "../Helpers/Misc.h"
+#include "../Helpers/StringParser.h"
 
 #ifdef USES_P023
 const char Plugin_023_myFont_Size[] PROGMEM = {
